# backend/swagger/dist/Tier0.openapi.yaml
openapi: 3.0.1
info:
  title: Tier0 UNS Open API
  description: UNS Open API for managing folders, files, templates, labels, and users
  version: 1.0.0
...
paths:
...
  /open-api/menu:
    post:
      summary: 保存菜单
      description: 保存菜单信息，支持图标上传
      operationId: saveMenu
      tags:
        - 菜单管理
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                serviceName:
                  type: string
                  description: 服务名（可选）
                name:
                  type: string
                  description: 名称，唯一标识一个菜单
                showName:
                  type: string
                  description: 显示名
                description:
                  type: string
                  description: 描述（可选）
                baseUrl:
                  type: string
                  description: baseURL
                openType:
                  type: integer
                  description: 跳转方式：0-iframe打开，1-打开新页面，2-app
                  enum: [0, 1, 2]
                file:
                  type: string
                  format: binary
                  description: Icon 文件
              required:
                - name
                - showName
                - baseUrl
                - openType
        required: true
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultVO'
      security:
        - apikey-header-apikey: []
components:
  schemas:
    ResultVO:
      type: object
      properties:
        code:
          type: integer
          format: int32
          description: 状态码 200-成功 400-参数错误 500-接口异常
        msg:
          type: string
          description: 响应消息
        data:
          type: object
          description: 响应数据
          additionalProperties: true
...
  securitySchemes:
    apikey-header-apikey:
      type: apiKey
      in: header
      name: apikey
      description: API Key认证
servers:
  - url: /
    description: 本地开发服务器
security: []